+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Information message shows amount of created and skipped items.
   - Button to view what items got skipped, if any got skipped.
   - Overview of skipped items and the reason for the skip when button gets clicked.
+- Pluralization in messages according to item count (e.g., "1 item", "2 items").
 
 ### Removed
 - Information message for each separate skipped file.